Create pop-up container for pop-up behavior.
